Add a pure nodejs http client to prevent adding curl to image
Docker HEALTHCHECKS are executed from within the container. To support standard docker healthchecks there need to be a script that is able to check if the server is running correctly. To avoid further dependancies like curl and to be able to use a distroless base in the future the checktool is implemented in JavaScript and executed with node like the server itself. Using the same technology provides the possiblity load the same config file and thus make the healthcheck in sync with the server configuration in regards of pathes and ports.